Render HTML in episode pop up
