+++ RocketCast/DefaultLibraryExtensions.swift
@@ -7,6 +7,7 @@
 //
 
 import Foundation
+import UIKit
 
 
 //from stackoverflow:  http://stackoverflow.com/questions/28059543/swift-replace-multiple-characters-in-string
@@ -17,6 +18,20 @@ extension String {
         return resultString
     }
     
+    // http://stackoverflow.com/questions/37048759/swift-display-html-data-in-a-label-or-textview
+    func stringFromHtml(string: String) -> NSAttributedString? {
+        do {
+            let data = string.data(using: String.Encoding.unicode, allowLossyConversion: true)
+            if let d = data {
+                let str = try NSAttributedString(data: d,
+                                                 options: [NSDocumentTypeDocumentAttribute: NSHTMLTextDocumentType],
+                                                 documentAttributes: nil)
+                return str
+            }
+        } catch {
+        }
+        return nil
+    }
 }
